ufs: export UFS headers to userspace
create UFS headers that can be exported to userspace.
